+++ nextstep-mvc/src/main/java/nextstep/mvc/DispatcherServlet.java
@@ -2,7 +2,7 @@ package nextstep.mvc;
 
 import nextstep.mvc.exception.AdapterNotFoundException;
 import nextstep.mvc.exception.HandlerNotFoundException;
-import nextstep.mvc.tobe.ModelAndView;
+import nextstep.mvc.tobe.view.ModelAndView;
 import nextstep.mvc.tobe.adapter.HandlerAdapter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
