+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/adapter/HandlerAdapter.java
@@ -1,11 +1,12 @@
 package nextstep.mvc.tobe.adapter;
 
-import nextstep.mvc.tobe.ModelAndView;
+import nextstep.mvc.tobe.view.ModelAndView;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 public interface HandlerAdapter {
     boolean isSupport(Object handler);
+
     ModelAndView handle(HttpServletRequest req, HttpServletResponse resp, Object handler) throws Exception;
 }
