+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/util/ComponentScanner.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc.tobe.util;
 
 import nextstep.web.annotation.Controller;
 import org.reflections.Reflections;
